full registry generated during get plugins, moved testing functionality back on class to occur when setting up full registry
